analizer/requirements.txt
pika
requests
beautifulsoup4
python-dotenv
